Pre-modification of consume spectra method
